Move the selection helpers to nab/_run.py so nab._lock imports alone
